Use present simple form for howto section
This ensures we consistently use the present simple form in the how-to
section. E.g:

  release instead of releasing
  create instead of creating
  etc.

This also renames the section itself from `how-tos` -> `how-to`.

The overall intent of these changes is to make the toctree and URLs read
like the action they are designed to teach. For example, the URL:

  /how-to/release/mozillavpn.html

is a clear indicator of what that page is going to be all about.