Remove empty 'architecture' section
It looks like the intent was to fill this out, but no one got around to
it. Let's remove it for now and re-add if anyone gets motivated enough
to start again in the future.

Though I think it makes sense for architecture docs to live next to the
repo that they are for, rather than in a centralized place like
relengdocs.