refactor!: fully switch to Diataxis system
BREAKING CHANGE: This will break pretty much every link to our docs.

Awhile back we agreed to try and implement the Diataxis documentation
system:
https://diataxis.fr/

However, we only moved a handful of docs over and have been in a mixed
state of having docs following two separate structures (if you can call
the old way a structure) ever since.

The goal of this PR is to re-organize all the docs in the old structured
over to the new Diataxis-based structure. There are plenty of docs that
don't fit neatly in one place or the other, so this PR often takes a
best guess at where it belongs.

This PR also doesn't split up any individual files, though it does
sometimes move docs from directories to different places.