Update the new ESR docs based on my experience with ESR140 (#310)
Updating gecko_taskgraph depends on balrog since you need the rule IDs.

You can run the esr staging early in the nightly cycle, it's unlikely
that something will break for ESRs during that cycle as it'll be on
people's mind, unlike the rest of the year. And this avoids having to
uplift patches later.

Likewise, the beta as esr release simulation should be done somewhat
early in the beta cycle, this will avoid potential uplifts to the esr
branch to fix said tests since they can be fixed during the beta cycle
instead.

I also added some command examples because they're not necessarily easy
to figure out and if someone else has to do it for the first time,
they'll be happy to find them here.

Added a new step required for the uplift approval mapping

Added a reference to ronin_puppet in the scriptworker bump. It's
something that I didn't think about when I was updating scriptworker
everywhere else.

Figuring out how to use the lando CLI was also a huge pain so I added
some documentation about that too.