Be a bit more explicit about waiting for VPN PPA packages to be published (#316)
We've had a full version not being published because we didn't wait long
enough after enabling publishing before disabling it again. The UI is
kind of confusing because the main page says that the package is
published even when it's not. The linked one has more detail and will
show `Pending` when a version isn't published so that's what we should
be checking.